+++ react-native.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  dependency: {
+    platforms: {
+      cmakeListsPath: 'cpp/CMakeLists.txt', 
+    },
+  },
+};